refactor: NoneDiscountPolicy의 트레이드오프 해결 위해 인터페이스 추가
