+++ chapter_2/src/main/java/org/movie/domain/discount/policy/AmountDiscountPolicy.java
@@ -4,7 +4,7 @@ import org.movie.domain.Money;
 import org.movie.domain.Screening;
 import org.movie.domain.discount.condition.DiscountCondition;
 
-public class AmountDiscountPolicy extends DiscountPolicy {
+public class AmountDiscountPolicy extends DefaultDiscountPolicy {
     private Money discountAmount;
 
     public AmountDiscountPolicy(Money discountAmount, DiscountCondition... conditions) {
