+++ chapter_2/src/main/java/org/movie/domain/discount/policy/DefaultDiscountPolicy.java
@@ -1,0 +1,28 @@
+package org.movie.domain.discount.policy;
+
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+import org.movie.domain.Money;
+import org.movie.domain.Screening;
+import org.movie.domain.discount.condition.DiscountCondition;
+
+public abstract class DefaultDiscountPolicy implements DiscountPolicy {
+    private List<DiscountCondition> conditions = new ArrayList<>();
+
+    public DefaultDiscountPolicy(DiscountCondition... conditions) {
+        this.conditions = Arrays.asList(conditions);
+    }
+
+    @Override
+    public Money calculateDiscountAmount(Screening screening) {
+        for (DiscountCondition condition : conditions) {
+            if (condition.isSatisfiedBy(screening)) {
+                return getDiscountAmount(screening);
+            }
+        }
+        return Money.ZERO;
+    }
+
+    abstract protected Money getDiscountAmount(Screening screening);
+}
