+++ chapter_2/src/main/java/org/movie/domain/discount/policy/DiscountPolicy.java
@@ -1,27 +1,8 @@
 package org.movie.domain.discount.policy;
 
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
 import org.movie.domain.Money;
 import org.movie.domain.Screening;
-import org.movie.domain.discount.condition.DiscountCondition;
 
-public abstract class DiscountPolicy {
-    private List<DiscountCondition> conditions = new ArrayList<>();
-
-    public DiscountPolicy(DiscountCondition... conditions) {
-        this.conditions = Arrays.asList(conditions);
-    }
-
-    public Money calculateDiscountAmount(Screening screening) {
-        for (DiscountCondition condition : conditions) {
-            if (condition.isSatisfiedBy(screening)) {
-                return getDiscountAmount(screening);
-            }
-        }
-        return Money.ZERO;
-    }
-
-    abstract protected Money getDiscountAmount(Screening screening);
+public interface DiscountPolicy {
+    Money calculateDiscountAmount(Screening screening);
 }
