+++ chapter_2/src/main/java/org/movie/domain/discount/policy/NoneDiscountPolicy.java
@@ -3,9 +3,10 @@ package org.movie.domain.discount.policy;
 import org.movie.domain.Money;
 import org.movie.domain.Screening;
 
-public class NoneDiscountPolicy extends DiscountPolicy {
+public class NoneDiscountPolicy implements DiscountPolicy {
+
     @Override
-    protected Money getDiscountAmount(Screening screening) {
+    public Money calculateDiscountAmount(Screening screening) {
         return Money.ZERO;
     }
 }
