+++ chapter_2/src/main/java/org/movie/domain/discount/policy/PercentDiscountPolicy.java
@@ -4,7 +4,7 @@ import org.movie.domain.Money;
 import org.movie.domain.Screening;
 import org.movie.domain.discount.condition.DiscountCondition;
 
-public class PercentDiscountPolicy extends DiscountPolicy{
+public class PercentDiscountPolicy extends DefaultDiscountPolicy {
     private double percent;
 
     public PercentDiscountPolicy(double percent, DiscountCondition... conditions) {
